update: new e2e tests files structure
