+++ e2e/testcases/e2e-cli-004_invalid_combination.go
@@ -1,0 +1,23 @@
+package testcases
+
+// E2E-CLI-004 - KICS has an invalid flag combination
+// an error message and return exit code 1
+
+func init() {
+	testSample := TestCase{
+		Name: "should display an error [E2E-CLI-004]",
+		Args: args{
+			Args: []cmdArgs{
+				[]string{"scan", "--ci", "--verbose"},
+				[]string{"--ci", "scan", "--verbose"},
+			},
+			ExpectedOut: []string{
+				"E2E_CLI_004",
+				"E2E_CLI_004",
+			},
+		},
+		WantStatus: []int{126, 126},
+	}
+
+	Tests = append(Tests, testSample)
+}
