+++ e2e/testcases/e2e-cli-008_scan_silent.go
@@ -1,0 +1,19 @@
+package testcases
+
+// E2E-CLI-008 - KICS  scan with --silent global flag
+// should hide all the output text in the CLI (empty output)
+
+func init() {
+	testSample := TestCase{
+		Name: "should hide all output text in CLI [E2E-CLI-008]",
+		Args: args{
+			Args: []cmdArgs{
+				[]string{"scan", "--silent", "-q", "../assets/queries", "-p", "fixtures/samples/terraform.tf"},
+			},
+			ExpectedOut: []string{"E2E_CLI_008"},
+		},
+		WantStatus: []int{50},
+	}
+
+	Tests = append(Tests, testSample)
+}
