+++ e2e/testcases/e2e-cli-012_scan_minimal-ui.go
@@ -1,0 +1,25 @@
+package testcases
+
+import "regexp"
+
+// E2E-CLI-012 - kics scan with minimal-ui flag should perform a scan
+// without showing detailed results on each line of code
+func init() {
+	testSample := TestCase{
+		Name: "should display minimal-ui [E2E-CLI-012]",
+		Args: args{
+			Args: []cmdArgs{
+				[]string{"scan", "-q", "../assets/queries", "-p", "../test/fixtures/tc-sim01/positive1.tf", "--minimal-ui"},
+			},
+		},
+		WantStatus: []int{50},
+		Validation: func(outputText string) bool {
+			match1, _ := regexp.MatchString("Description:", outputText)
+			match2, _ := regexp.MatchString("Platform:", outputText)
+			// if not found -> the the test was successful
+			return !match1 && !match2
+		},
+	}
+
+	Tests = append(Tests, testSample)
+}
