+++ e2e/testcases/e2e-cli-018_scan_exclude-categories.go
@@ -1,0 +1,21 @@
+package testcases
+
+// E2E-CLI-018  - KICS scan command with --exclude-categories flag
+// should not run queries that are part of the provided categories.
+func init() {
+	testSample := TestCase{
+		Name: "should exclude provided categories [E2E-CLI-018]",
+		Args: args{
+			Args: []cmdArgs{
+				[]string{"scan", "--exclude-categories", "Observability,Insecure Configurations", "-s",
+					"-q", "../assets/queries", "-p", "fixtures/samples/terraform-single.tf"},
+
+				[]string{"scan", "-s",
+					"-q", "../assets/queries", "-p", "fixtures/samples/terraform-single.tf"},
+			},
+		},
+		WantStatus: []int{20, 40},
+	}
+
+	Tests = append(Tests, testSample)
+}
