+++ e2e/testcases/e2e-cli-021_scan_status_code.go
@@ -1,0 +1,24 @@
+package testcases
+
+// E2E-CLI-021 - KICS can return different status code based in the scan results (High/Medium/Low..)
+// when excluding categories/queries and losing results we can get a different status code.
+func init() {
+	testSample := TestCase{
+		Name: "should validate the kics status code [E2E-CLI-021]",
+		Args: args{
+			Args: []cmdArgs{
+				[]string{"scan",
+					"-q", "../assets/queries", "-p", "../test/fixtures/all_auth_users_get_read_access/test/positive.tf"},
+
+				[]string{"scan", "--exclude-categories",
+					"Access Control,Availability,Backup,Best Practices,Build Process,Encryption," +
+						"Insecure Configurations,Insecure Defaults,Networking and Firewall,Observability," +
+						"Resource Management,Secret Management,Supply-Chain,Structure and Semantics",
+					"-q", "../assets/queries", "-p", "../test/fixtures/all_auth_users_get_read_access/test/positive.tf"},
+			},
+		},
+		WantStatus: []int{50, 0},
+	}
+
+	Tests = append(Tests, testSample)
+}
