+++ e2e/testcases/e2e-cli-025_scan_fail-on.go
@@ -1,0 +1,21 @@
+package testcases
+
+// E2E-CLI-025 - KICS scan command with --fail-on flag should
+// return status code different from 0 only when results match the severity provided in this flag
+func init() {
+	testSample := TestCase{
+		Name: "should fail-on provided values [E2E-CLI-025]",
+		Args: args{
+			Args: []cmdArgs{
+				[]string{"scan", "--fail-on", "info,low",
+					"-s", "-q", "../assets/queries", "-p", "../assets/queries/dockerfile/apk_add_using_local_cache_path/test/positive.dockerfile"},
+
+				[]string{"scan", "--fail-on", "info",
+					"-s", "-q", "../assets/queries", "-p", "../assets/queries/dockerfile/apk_add_using_local_cache_path/test/positive.dockerfile"},
+			},
+		},
+		WantStatus: []int{30, 20},
+	}
+
+	Tests = append(Tests, testSample)
+}
