+++ e2e/testcases/e2e-cli-026_scan_ignore-on-exit.go
@@ -1,0 +1,28 @@
+package testcases
+
+// E2E-CLI-026 - KICS scan command with --ignore-on-exit flag
+// should return status code 0 if the provided flag occurs.
+// Example: '--ignore-on-exit errors' -> Returns 0 if an error was found, instead of 126/130...
+func init() {
+	testSample := TestCase{
+		Name: "should ignore on exit provided flags [E2E-CLI-026]",
+		Args: args{
+			Args: []cmdArgs{
+				[]string{"scan", "--ignore-on-exit",
+					"-s", "-q", "../assets/queries", "-p", "fixtures/samples/terraform-single.invalid.name"},
+
+				[]string{"scan", "--ignore-on-exit", "errors",
+					"-s", "-q", "../assets/queries", "-p", "fixtures/samples/terraform-single.invalid.name"},
+
+				[]string{"scan", "--ignore-on-exit", "errors",
+					"-s", "-q", "../assets/queries", "-p", "fixtures/samples/terraform-single.tf"},
+
+				[]string{"scan", "--ignore-on-exit", "all",
+					"-s", "-q", "../assets/queries", "-p", "fixtures/samples/terraform-single.tf"},
+			},
+		},
+		WantStatus: []int{126, 0, 40, 0},
+	}
+
+	Tests = append(Tests, testSample)
+}
