+++ e2e/testcases/e2e-cli-041_scan_remote_path_git.go
@@ -1,0 +1,25 @@
+package testcases
+
+// E2E-CLI-041 - Kics scan command with -p targeting remote path (git)
+// should download and scan the provided path.
+func init() {
+	testSample := TestCase{
+		Name: "should download and scan the provided git path [E2E-CLI-041]",
+		Args: args{
+			Args: []cmdArgs{
+				[]string{"scan", "--output-path", "output", "--output-name", "E2E_CLI_041_RESULT",
+					"--report-formats", "json,sarif,glsast", "-q", "../assets/queries",
+					"-p", "git::https://github.com/dockersamples/example-voting-app"},
+			},
+			ExpectedResult: []ResultsValidation{
+				{
+					ResultsFile:    "E2E_CLI_041_RESULT",
+					ResultsFormats: []string{"json", "sarif", "glsast"},
+				},
+			},
+		},
+		WantStatus: []int{50},
+	}
+
+	Tests = append(Tests, testSample)
+}
